vision_encoder/requirements.txt
torch==2.4.0
torchvision==0.19.0
timm==1.0.7
matplotlib==3.9.1
Pillow==10.4.0